+++ src/components/pages/SignUp.jsx
@@ -4,8 +4,8 @@ import { Link } from 'react-router-dom'
 
 import { Logo } from 'Images'
 
-import { Password } from 'components/molecules/Password'
-import { PasswordRepeat } from 'components/molecules/PasswordRepeat'
+import { Password } from 'components/shared/Password'
+import { PasswordRepeat } from 'components/shared/PasswordRepeat'
 import { eventToValue } from 'helpers/eventToValue'
 import { ApiContext } from 'providers/ApiProvider'
 import { SessionContext } from 'providers/SessionProvider'
